Raid WW2 & Payday all_h support.
Thanks Frankelstner for the payday 2 fix & spy for raid ww2 one, this saved me time from needing to research the binary.
pdmod for files in all_x bundles are not supported atm, you might get an error if you try patching it.
I dunno if I should even support it considering the chance that it will break from each update is almost a 100% and the fact that superblt/beardlib can replace most files in a better manner.